refactor: extract composer preparation from AgentWorkspace
